Module 5: moved proptypes
