+++ src/components/DialogOpenButton.tsx
@@ -2,7 +2,15 @@ import React, { useState, MouseEvent } from "react";
 import { createPortal } from 'react-dom';
 import Dialog from './Dialog';
 import MovieForm from './MovieForm';
-import {DialogOpenButtonProps} from '../types'
+import {MovieData} from '../types'
+
+export interface DialogOpenButtonProps {
+    movieData?: MovieData;
+    onDialogFormSubmit: () => void;
+    isDeleteForm?: boolean;
+    dialogTitle: string;
+    buttonText: string;
+}
 
 const DialogOpenButton: React.FC<DialogOpenButtonProps> = ({movieData, onDialogFormSubmit, isDeleteForm = false, dialogTitle, buttonText = "Open Dialog"}) => {
     const [isDialogOpen, setDialogOpen] = useState(false)
